git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@6883 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/operator/matrix/spmatrix.hh
@@ -646,7 +646,7 @@ private:
       for( DofIteratorType dofIt = diag.dbegin(); 
            dofIt != dofEnd; ++ dofIt, ++row ) 
       {
-        assert( row < matrix_.rows() );
+        assert( row < ( unsigned int )matrix_.rows() );
         (*dofIt) = matrix_( row, row );
       }
     }
